+++ main_project/component/component.hpp
@@ -49,9 +49,9 @@ public:
     // these are used for non linear analysis
     void setNodalVoltages(vector<float> v);
     // get the IV characteristics of a component given two nodes
-    virtual float ivAtNode(int n1) const =0;
+    virtual double ivAtNode(int n1) const =0;
     // get the derivative of the IV charateristic of a component given two nodes and the node the derivative is based on
-    virtual float divAtNode(int n1, int dn) const =0;
+    virtual double divAtNode(int n1, int dn) const =0;
 
     // this should be used to update the value of the voltage and current accross a component after an iteration
     virtual void updateVals(float newVoltage, float newCurrent, int order);
