+++ main_project/component/currentControlledCurrentSource.hpp
@@ -11,8 +11,8 @@ public:
     CurrentControlledCurrentSource(string _name, vector<string> args, vector<float> extraInfo);
     ~CurrentControlledCurrentSource() = default;
 
-    float ivAtNode(int n) const override;
-    float divAtNode(int n, int dn) const override;
+    double ivAtNode(int n) const override;
+    double divAtNode(int n, int dn) const override;
 
     float getTotalCurrent(const VectorXd &x, int highestNodeNumber, float voltage = 0, int order = 1)  override;
 };
