+++ main_project/component/currentSource.hpp
@@ -24,8 +24,8 @@ public:
 
     void updateVals(float time);
 
-    float ivAtNode(int n1) const override;
-    float divAtNode(int n1, int dn) const override;
+    double ivAtNode(int n1) const override;
+    double divAtNode(int n1, int dn) const override;
 
     float getCurrent() const override;
