+++ main_project/component/diode.hpp
@@ -16,8 +16,8 @@ public:
     Diode(string _name, float r, int n1, int n2);
     ~Diode() = default;
 
-    float ivAtNode(int n1) const override;
-    float divAtNode(int n1, int dn) const override;
+    double ivAtNode(int n1) const override;
+    double divAtNode(int n1, int dn) const override;
 
     vector<int> getNodes() const override;
