+++ main_project/component/exampleInputDiode.hpp
@@ -29,8 +29,8 @@ public:
     vector<int> getNodes() const override;
     string getModelName() const override;
 
-    float ivAtNode(int n1) const override;
-    float divAtNode(int n1, int dn) const override;
+    double ivAtNode(int n1) const override;
+    double divAtNode(int n1, int dn) const override;
 };
 
 #endif
