+++ main_project/component/inductor.hpp
@@ -20,8 +20,8 @@ public:
 	float getConductance() const override;
 	vector<int> getNodes() const ;
 
-	float ivAtNode(int n) const override;
-    float divAtNode(int n, int dn) const override;
+	double ivAtNode(int n) const override;
+    double divAtNode(int n, int dn) const override;
 
 	float getCurrent() const override;
 	float getTotalCurrent(const VectorXd &x, int highestNodeNumber, float voltage = 0, int order = 1)  override;
