+++ main_project/component/mosfet.hpp
@@ -28,8 +28,8 @@ public:
     void addParam(int paramId, float paramValue) override;
     string getModelName() const override;
 
-    float ivAtNode(int n1) const override;
-    float divAtNode(int n1, int dn) const override;
+    double ivAtNode(int n1) const override;
+    double divAtNode(int n1, int dn) const override;
 
     vector<int> getNodes() const override;
